+++ packages/contracts/source/contracts/consensus/index.ts
@@ -1,0 +1,1 @@
+export * from "./consensus";
